refactor(tests)/separate scripts for api and app tests
