+++ apps/e2e/tests/app/404.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { NotFoundPage } from '../actions/pages/NotFoundPage';
+import { NotFoundPage } from '../../actions/pages/NotFoundPage';
 
 test('redirects to the not found page, redirects to the home page by click on the link', async ({
 	page,
